- Loggin package fixed - Moved properties to parent pom
