+++ pom.xml
@@ -17,6 +17,7 @@
 		<spring.boot.version>2.4.1</spring.boot.version>
 		<kotlin.version>1.4.21</kotlin.version>
 		<spring-cloud-gcp.version>1.2.5.RELEASE</spring-cloud-gcp.version>
+		<kotlin-coroutines.version>1.3.8</kotlin-coroutines.version>
 	</properties>
 
 	<modules>
